requirements.txt
# requirements.txt
streamlit
sentence-transformers
faiss-cpu
python-pptx
python-docx
PyMuPDF
ollama